refactor(shared-docs): move the testing helpers into a public location
Make the testing helpers available at @angular/docs/testing
