+++ docs/components/cookie-popup/cookie-popup.component.spec.ts
@@ -10,7 +10,7 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {CookiePopup, STORAGE_KEY} from './cookie-popup.component';
 import {LOCAL_STORAGE} from '../../providers/index';
-import {MockLocalStorage} from '../../utils/testing-helper';
+import {MockLocalStorage} from '../../testing/index';
 
 describe('CookiePopup', () => {
   let fixture: ComponentFixture<CookiePopup>;
