+++ docs/components/search-dialog/BUILD.bazel
@@ -42,7 +42,7 @@ ts_library(
         "//docs/components/algolia-icon",
         "//docs/providers",
         "//docs/services",
-        "//docs/utils:testing-helper",
+        "//docs/testing",
         "@npm//@angular/core",
         "@npm//@angular/platform-browser",
         "@npm//@angular/router",
